pyproject.toml
[project]
name = "ai-marketing"
version = "0.1.0"
description = "AI-powered marketing automation for Pickleball Effect"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-bigquery>=3.25",
    "google-auth>=2.29",
    "requests>=2.31",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "anthropic>=0.40",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["ingestion"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
